requirements.txt
Package                       Version
----------------------------- --------
asgiref                       3.7.2
astroid                       3.0.3
atomicwrites                  1.4.1
attrs                         23.2.0
certifi                       2024.2.2
charset-normalizer            2.0.12
colorama                      0.4.6
dill                          0.3.8
Django                        3.2
django-filter                 23.5
djangorestframework           3.12.4
djangorestframework-simplejwt 5.3.1
flake8                        7.0.0
idna                          3.6
iniconfig                     2.0.0
isort                         5.13.2
mccabe                        0.7.0
packaging                     23.2
pip                           24.0
platformdirs                  4.2.0
pluggy                        0.13.1
py                            1.11.0
pycodestyle                   2.11.1
pyflakes                      3.2.0
PyJWT                         2.1.0
pylint                        3.0.3
pytest                        6.2.4
pytest-django                 4.4.0
pytest-pythonpath             0.7.3
python-dotenv                 1.0.1
pytz                          2024.1
requests                      2.26.0
setuptools                    58.1.0
sqlparse                      0.4.4
toml                          0.10.2
tomli                         2.0.1
tomlkit                       0.12.3
typing_extensions             4.9.0
urllib3                       1.26.18
